fix(tests): simplify size comparison in test
Remove unnecessary static_cast in EXPECT_EQ comparison

+++ utils_test.cpp
@@ -62,7 +62,7 @@ TEST(InitDefaultHttpMethodsTests, InsertsFiveStandardMethods) {
     std::set<http::Method> methods;
     initDefaultHttpMethods(methods);
 
-    EXPECT_EQ(static_cast<size_t>(5), methods.size());
+    EXPECT_EQ(5u, methods.size());
 
     EXPECT_NE(methods.find(http::GET), methods.end());
     EXPECT_NE(methods.find(http::POST), methods.end());
